Created ToggleSwitch as separate component
