+++ src/features/common/InputTypes/ToggleSwitch.tsx
@@ -1,0 +1,18 @@
+import { withStyles } from '@material-ui/core/styles';
+import Switch from '@material-ui/core/Switch';
+
+const ToggleSwitch = withStyles({
+  switchBase: {
+    color: '#fff',
+    '&$checked': {
+      color: '#89B53A',
+    },
+    '&$checked + $track': {
+      backgroundColor: '#89B53A',
+    },
+  },
+  checked: {},
+  track: {},
+})(Switch);
+
+export default ToggleSwitch;
